feat: remove rahasia bengcare
